+++ Assets/Packages/com.studio23.ss2.loadingsystem/CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+
+## [v1.1.3] - 2024-09-17
+
+### Fixed
+
+- Fixed loading scene not instantiating loading prefab
+- Changed random algorithm
+
+
 ## [v1.1.2] - 2024-09-17
 
 ### Fixed